Fix a row-indexing bug with sparse matrices that have non-Int indices

+++ test/perf/shootout/fasta.jl
@@ -22,7 +22,7 @@ const IA  =   3877.0
 const IC  =  29573.0
 
 function gen_random()
-    global rng_state::Float64 = ((rng_state::Float64 * IA + IC) % IM) / IM
+    global rng_state = ((rng_state::Float64 * IA + IC) % IM) / IM
 end
 function repeat_fasta(src, n)
     k = length(src)
@@ -55,7 +55,7 @@ end
 
 rng_state = 42.0
 function fasta(n=25000000)
-  repeat_fasta(alu, 2n)
-  random_fasta(iub1, iub2, 3n)
-  random_fasta(homosapiens1, homosapiens2, 5n)
+    repeat_fasta(alu, 2n)
+    random_fasta(iub1, iub2, 3n)
+    random_fasta(homosapiens1, homosapiens2, 5n)
 end
